backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
uvloop==0.19.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
redis[hiredis]==5.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
cryptography==42.0.5
minio==7.2.5
weasyprint==61.2
openpyxl==3.1.2
celery==5.3.6
prometheus-fastapi-instrumentator==6.1.0
